Final Debug
In Employeewurde die ID nicht angezeigt

der Getter hieß "getemployID" muss aber  "getEmployID" heißen

+++ src/main/java/de/hitec/nhplus/controller/AllPatientController.java
@@ -31,7 +31,7 @@ public class AllPatientController extends LockingObjects{
     private TableView<Patient> tableView;
 
     @FXML
-    private TableColumn<Patient, Integer> columnId;
+    private TableColumn<Patient, Long> columnId;
 
     @FXML
     private TableColumn<Patient, String> columnFirstName;
